requirements.txt
streamlit
openai
mysql-connector-python
st_pages
pandas
